@@ -28,7 +28,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
+        "php": ">=8.1",
         "yiisoft/yii2": "^2.0.15",
         "simialbi/yii2-simialbi-base": ">=0.13.1 <1.0 | ^1.0.0"
     },
